requirements.txt
amqp==5.0.9
asgiref==3.5.0
attrs==21.4.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.0.2
django-background-tasks==1.2.5
django-compat==1.0.15
django-filter==21.1
djangorestframework==3.13.1
future==0.18.2
idna==3.3
importlib-resources==5.4.0
jsonschema==4.4.0
kombu==5.2.3
pkg_resources==0.0.0
prompt-toolkit==3.0.28
pyrsistent==0.18.1
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
serializers==0.2.4
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
zipp==3.7.0
